memory_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pgvector==0.2.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
google-generativeai==0.3.2
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6